close #524
added documentation. @strengejacke this function is not used for now anywhere in easystats, but it was made in the context of the unified effect size idea that we discussed a long time ago (but no advances have been on that front for now, so :shrug:). If this function bothers you I suppose we could put it in a WIP folder, I don't think it would lead to breaking changes.